divide branch scraper
