+++ lib/branch_scraper.rb
@@ -1,0 +1,67 @@
+require_relative 'scraper'
+
+class BranchScraper < Scraper
+
+  def get_all_branches(bank_code)
+    search_bank(bank_code)
+    siten_list_pages = get_branch_list_pages
+    branches = get_all_branches_from_list_pages(siten_list_pages)
+    branches.flatten
+  end
+
+  def search_bank(bank_code)
+    agent.page.link_with( :text => "コードから検索<入力形式>" ).click
+    form = agent.page.forms[0]
+    form.inbc = bank_code
+    agent.submit(form)
+    form = agent.page.forms[1]
+    agent.submit(form)
+  end
+
+  def get_branch_list_pages
+    siten_list_pages = []
+    kanas.each do |kana|
+      siten_list_page = click_siten_kana_link(kana)
+      display_page_info(siten_list_page)
+      siten_list_pages << siten_list_page
+    end
+    siten_list_pages
+  end
+
+  def get_all_branches_from_list_pages(siten_list_pages)
+    branches = []
+    siten_list_pages.each do |list_page|
+      kana_branches = get_branches_from_list(list_page)
+      branches << kana_branches
+    end
+    branches.delete([])
+    branches
+  end
+
+  def get_branches_from_list(branch_list_page)
+    branches = []
+    tablerows = branch_list_page.search('table.tbl1 tr')
+    tablerows.shift
+    tablerows.each do |tr|
+      name = tr.css('td.g1:first-child').inner_text
+      if name == "該当するデータはありません"
+        next
+      end
+      code = tr.css('td.g2').inner_text
+      branch = Branch.new(code, name)
+      branches << branch
+    end
+    branches
+  end
+
+  def click_siten_kana_link(kana)
+    kana_form = agent.page.forms[1]
+    button = kana_form.button_with(:value => kana)
+    list_page = agent.submit(kana_form, button)
+    back_form = agent.page.forms[0]
+    back_button = back_form.button_with(:value => "前ページに戻る")
+    agent.submit(back_form, back_button)
+    list_page
+  end
+
+end
